services/certificate-stream/requirements.txt
certstream==1.12
psycopg2-binary==2.9.9
